Move telemetry helpers within their plugin
