+++ packages/core/package.json
@@ -24,14 +24,10 @@
     },
     "dependencies": {
         "chalk": "2.3.1",
-        "esbuild": "0.20.2",
-        "fs-extra": "7.0.1",
-        "pretty-bytes": "5.6.0",
         "unplugin": "1.10.1"
     },
     "devDependencies": {
         "@types/chalk": "2.2.0",
-        "@types/fs-extra": "8.1.0",
         "@types/node": "18.15.3",
         "typescript": "5.4.3"
     }
