+++ packages/plugins/telemetry/package.json
@@ -27,12 +27,14 @@
         "@dd/core": "workspace:*",
         "chalk": "2.3.1",
         "esbuild": "0.20.2",
+        "fs-extra": "7.0.1",
         "pretty-bytes": "5.6.0",
         "unplugin": "1.10.1",
         "webpack": "5.49.0"
     },
     "devDependencies": {
-        "@types/chalk": "2.2.0"
+        "@types/chalk": "2.2.0",
+        "@types/fs-extra": "8.1.0"
     },
     "peerDependencies": {
         "esbuild": ">=0.x",
