+++ packages/plugins/telemetry/src/common/aggregator.ts
@@ -2,9 +2,15 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import type { Report, StatsJson, BundlerStats, EsbuildStats } from '@dd/core/types';
-
-import type { Metric, MetricToSend, OptionsDD } from '../types';
+import type {
+    Report,
+    StatsJson,
+    BundlerStats,
+    EsbuildStats,
+    Metric,
+    MetricToSend,
+    OptionsDD,
+} from '../types';
 
 import { getMetric } from './helpers';
 import {
