+++ packages/plugins/telemetry/src/common/filters.ts
@@ -1,3 +1,7 @@
+// Unless explicitly stated otherwise all files in this repository are licensed under the MIT License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2019-Present Datadog, Inc.
+
 import type { Metric } from '../types';
 
 const filterTreeMetrics = (metric: Metric): Metric | null =>
