+++ packages/plugins/telemetry/src/common/metrics/common.ts
@@ -2,9 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import type { Report, LocalModule, TimingsMap, BundlerStats } from '@dd/core/types';
-
-import type { Metric } from '../../types';
+import type { Report, LocalModule, TimingsMap, BundlerStats, Metric } from '../../types';
 import { flattened, getType } from '../helpers';
 
 interface GeneralReport {
