+++ packages/plugins/telemetry/src/common/modules.ts
@@ -2,9 +2,11 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { formatModuleName, getDisplayName } from '@dd/core/helpers';
-import type { LocalModule } from '@dd/core/types';
 import type { Metafile } from 'esbuild';
+
+import type { LocalModule } from '../types';
+
+import { formatModuleName, getDisplayName } from './helpers';
 
 const modulesMap: { [key: string]: LocalModule } = {};
