+++ packages/plugins/telemetry/src/webpack-plugin/loaders.ts
@@ -2,11 +2,10 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { getDisplayName, getModuleName, getLoaderNames } from '@dd/core/helpers';
-import type { Module, Event, Timing, Compilation, TimingsMap } from '@dd/core/types';
 import { performance } from 'perf_hooks';
 
-import type { TelemetryOptions } from '../types';
+import { getDisplayName, getModuleName, getLoaderNames } from '../common/helpers';
+import type { Module, Event, Timing, Compilation, TimingsMap, TelemetryOptions } from '../types';
 
 export class Loaders {
     constructor(cwd: string, options: TelemetryOptions) {
