+++ packages/tests/src/plugins/telemetry/webpack-plugin/modules.test.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import type { LocalModule, Module, Compilation, Chunk } from '@dd/core/types';
+import type { LocalModule, Module, Compilation, Chunk } from '@dd/telemetry-plugins/types';
 import { Modules } from '@dd/telemetry-plugins/webpack-plugin/modules';
 import { mockTelemetryOptions } from '@dd/tests/plugins/telemetry/testHelpers';
